Get rate from local database

+++ app/src/main/java/aposalo/com/currencycalculator/domain/server/dto/Info.kt
@@ -1,6 +1,6 @@
 package aposalo.com.currencycalculator.domain.server.dto
 
 data class Info (
-    val rate: Double,
+    val rate: Float,
     val timestamp: Int
 )
